migrated viscosity function
